Have array helpers responsible for ufunc behaviour
