+++ hypothesis-python/tests/numpy/test_gufunc.py
@@ -20,6 +20,10 @@ from pytest import param
 from hypothesis import example, given, note, settings, strategies as st
 from hypothesis.errors import InvalidArgument
 from hypothesis.extra import numpy as nps
+from hypothesis.extra._array_helpers import (
+    _SIGNATURE,
+    _hypothesis_parse_gufunc_signature,
+)
 
 from tests.common.debug import find_any, minimal
 
@@ -55,7 +59,7 @@ def test_numpy_signature_parses(sig):
 
     np_sig = np.lib.function_base._parse_gufunc_signature(sig)
     try:
-        hy_sig = nps._hypothesis_parse_gufunc_signature(sig, all_checks=False)
+        hy_sig = _hypothesis_parse_gufunc_signature(sig, all_checks=False)
         assert np_sig == hy_sig_2_np_sig(hy_sig)
     except InvalidArgument:
         shape_too_long = any(len(s) > 32 for s in np_sig[0] + np_sig[1])
@@ -67,14 +71,14 @@ def test_numpy_signature_parses(sig):
         sig = in_ + "->" + out.split(",(")[0]
         np_sig = np.lib.function_base._parse_gufunc_signature(sig)
         if all(len(s) <= 32 for s in np_sig[0] + np_sig[1]):
-            hy_sig = nps._hypothesis_parse_gufunc_signature(sig, all_checks=False)
+            hy_sig = _hypothesis_parse_gufunc_signature(sig, all_checks=False)
             assert np_sig == hy_sig_2_np_sig(hy_sig)
 
 
 @use_signature_examples
-@given(st.from_regex(nps._SIGNATURE))
+@given(st.from_regex(_SIGNATURE))
 def test_hypothesis_signature_parses(sig):
-    hy_sig = nps._hypothesis_parse_gufunc_signature(sig, all_checks=False)
+    hy_sig = _hypothesis_parse_gufunc_signature(sig, all_checks=False)
     try:
         np_sig = np.lib.function_base._parse_gufunc_signature(sig)
         assert np_sig == hy_sig_2_np_sig(hy_sig)
@@ -82,13 +86,13 @@ def test_hypothesis_signature_parses(sig):
         assert "?" in sig
         # We can always fix this up, and it should then always validate.
         sig = sig.replace("?", "")
-        hy_sig = nps._hypothesis_parse_gufunc_signature(sig, all_checks=False)
+        hy_sig = _hypothesis_parse_gufunc_signature(sig, all_checks=False)
         np_sig = np.lib.function_base._parse_gufunc_signature(sig)
         assert np_sig == hy_sig_2_np_sig(hy_sig)
 
 
 def test_frozen_dims_signature():
-    nps._hypothesis_parse_gufunc_signature("(2),(3)->(4)")
+    _hypothesis_parse_gufunc_signature("(2),(3)->(4)")
 
 
 @st.composite
@@ -183,7 +187,7 @@ def gufunc_sig_to_einsum_sig(gufunc_sig):
         assert "x" not in labels, "we reserve x for fixed-dimensions"
         return "..." + "".join(i if not i.isdigit() else "x" for i in labels)
 
-    gufunc_sig = nps._hypothesis_parse_gufunc_signature(gufunc_sig)
+    gufunc_sig = _hypothesis_parse_gufunc_signature(gufunc_sig)
     input_sig = ",".join(map(einlabels, gufunc_sig.input_shapes))
     return input_sig + "->" + einlabels(gufunc_sig.result_shape)
